Refactor: 기존 코드 [분석, 신체정보/목표영양성분] 삭제
