--- MeokPT/Sources/Analyze/Extension/Notification.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-extension Notification.Name {
-    static let didUpdateNutritionItems = Notification.Name("didUpdateNutritionItems")
-    static let didUpdateDietItems = Notification.Name("didUpdateDietItems")
-}
