--- MeokPT/Sources/Analyze/View/Conditional/Diet/DietNotEmptyView.swift
@@ -1,24 +1,0 @@
-import SwiftUI
-
-struct DietNotEmptyView: View {
-    let dietItems: [DietItem]
-    let onMealTypeChange: (DietItem.ID, MealType) -> Void
-
-    var body: some View {
-        ScrollView {
-            VStack(spacing: 20) {
-                ForEach(dietItems) { item in
-                    DietItemView(item: item, onMealTypeChange: onMealTypeChange)
-                }
-            }
-        }
-    }
-}
-
-#Preview {
-    DietNotEmptyView(
-        dietItems: mockDietItemsForPreview,
-        onMealTypeChange: { _, _ in }
-    )
-}
-
