--- MeokPT/Sources/Analyze/View/Conditional/Diet/DietNutritionInfoCellView.swift
@@ -1,17 +1,0 @@
-import SwiftUI
-
-struct DietNutritionInfoCellView: View {
-    let name: String
-    let value: String
-    
-    var body: some View {
-        VStack(alignment: .center) {
-            Text(name)
-                .font(.caption)
-                .foregroundStyle(Color("AppSecondaryColor"))
-            Spacer().frame(height: 4)
-            Text(value)
-                .font(.body)
-        }
-    }
-}
