--- MeokPT/Sources/MyPage/MyData/Model/Gender.swift
@@ -1,8 +1,0 @@
-enum Gender: String, CaseIterable, Identifiable, Equatable, Codable, Hashable {
-    case female = "여성"
-    case male = "남성"
-    
-    var id: String { rawValue }
-    
-    var description: String { rawValue }
-}
