--- MeokPT/Sources/MyPage/MyData/Model/NutritionRowData.swift
@@ -1,8 +1,0 @@
-struct NutritionRowData: Identifiable, Equatable {
-    let type: NutritionType
-    var value: String
-    
-    var id: String { type.id }
-    var name: String { type.rawValue }
-    var unit: String { type.unit }
-}
