+++ MeokPT/Sources/MyPage/MyPageView.swift
@@ -8,7 +8,6 @@ struct MyPageView: View {
     var body: some View {
         NavigationStack {
             List {
-                // MARK: - Profile Section
                 Section {
                     Button {
                         store.send(store.userProfile == nil ? .loginSignUpButtonTapped : .profileEditButtonTapped)
@@ -41,38 +40,25 @@ struct MyPageView: View {
                     }
                 }
 
-                // MARK: - 주요 기능
                 Section(header: Text("분석용 정보 입력")) {
                     NavigationLink {
-                        BodyNutritionContainerView(
-                            initialTab: .bodyinInfoInput,
-                            bodyInfoStore: Store(initialState: BodyInfoInputFeature.State()) { BodyInfoInputFeature() },
-                            nutritionStore: Store(initialState: DailyNutritionFeature.State()) { DailyNutritionFeature() }
-                        )
                     } label: {
                         Label("신체정보 입력", systemImage: "person.fill")
                     }
 
                     NavigationLink {
-                        BodyNutritionContainerView(
-                            initialTab: .dailyNutrition,
-                            bodyInfoStore: Store(initialState: BodyInfoInputFeature.State()) { BodyInfoInputFeature() },
-                            nutritionStore: Store(initialState: DailyNutritionFeature.State()) { DailyNutritionFeature() }
-                        )
                     } label: {
                         Label("하루 섭취량 입력", systemImage: "fork.knife")
                     }
                 }
-                
                 Section(header: Text("저장한 분석")) {
                     NavigationLink {
-                        AIHistoryView(store: Store(initialState: AIHistoryFeature.State()) { AIHistoryFeature() })
+//                        AIHistoryView(store: Store(initialState: AIHistoryFeature.State()) { AIHistoryFeature() })
                     } label: {
                         Label("저장한 분석 리스트 보기", systemImage: "list.bullet.clipboard.fill")
                     }
                 }
 
-                // MARK: - 내 활동
                 if store.currentUser != nil {
                     Section(header: Text("내 활동")) {
                         NavigationLink {
@@ -82,7 +68,6 @@ struct MyPageView: View {
                         }
                     }
 
-                    // MARK: - 계정 설정
                     Section(header: Text("계정")) {
                         Button {
                             store.showLogoutAlert = true
